Make it easier to work with stream position range
